+++ src/main/kotlin/domain/stone/BlackStones.kt
@@ -1,3 +1,3 @@
-package domain
+package domain.stone
 
 class BlackStones private constructor(stones: List<Stone> = emptyList()) : Stones(stones)
