+++ src/main/kotlin/domain/stone/StoneType.kt
@@ -1,4 +1,4 @@
-package domain
+package domain.stone
 
 enum class StoneType {
     BLACK, WHITE
