+++ src/main/kotlin/domain/stone/WhiteStones.kt
@@ -1,3 +1,3 @@
-package domain
+package domain.stone
 
 class WhiteStones private constructor(stones: List<Stone> = emptyList()) : Stones(stones)
