+++ src/test/kotlin/domain/stone/StonePositionTest.kt
@@ -1,4 +1,4 @@
-package domain
+package domain.stone
 
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Assertions.assertNull
